requirements.txt
mpxpy
python-dotenv
python-pptx
requests
